CrunchbaseData.SQL
Crunchbase data is a crowdsourced index of startups, founders, investors, and activities. The specific data was collected February 5th, 2014. This exploration consists of multiple joins and clauses to look at various data on investors for the startups and which are acquired and are not. Lastly, UNIONs are used between 2 datasets.